Update local/IMCN - rcpch-growth-chart.t.json
Each observation is to be unique. However, should it be incorporated within the payload response? If so, is there a need to send the individual observation if the app controls the logic? 
Do we force the obs data in first, then a retrieval? Would  that be too API chatty? 
Do we use the “URI to original measurement” properly, as per template instruction?
Query to plot current child growth chart data to be an AQL call via the API. Need to build this. 
Query: Use SCT codes in the header?
Change the Indicator.Comment to Centile band
Relabelled bone age.confounding factors to events text.